requirements.txt
# Core Python libraries (built-in, no installation needed)
# os, sys, datetime, json, hashlib, re, io, traceback, logging

# Third-party libraries
google-api-python-client==2.97.0
google-auth-oauthlib==1.2.0
google-auth==2.22.0
python-docx==1.1.0
PyMuPDF>=1.21.0
pdfminer.six>=20221105
pdf2image>=1.16.0
pytesseract>=0.3.9
psutil>=6.0.0
pandas>=2.2.3
openpyxl>=3.1.2
xlrd>=2.0.1
